simpletix/requirements-dev.txt
# requirements-dev.txt
# Include runtime dependencies
-r requirements.txt

# Testing and development dependencies
black==24.8.0
flake8==7.1.1
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
coverage>=7,<8
coveralls==4.0.1 ; python_version < "3.13"

